requirements.txt
ipykernel
langchain_google_genai
langchain
flask
sentence-transformers
pypdf
python-dotenv
langchain-pinecone
langchain-community
langchain_huggingface
-e .